requirements.txt
numpy>=1.24.0
dropbox>=11.36.2
hypothesis>=6.98.0
pytest>=7.4.0
jsonschema>=4.0.0
PyYAML>=6.0.1
gmsh==4.11.1
scikit-learn>=1.2.0




